:rocket: finished expanser
*
- norm check
