+++ Makefile
@@ -24,7 +24,7 @@ INC_DIR		= inc/
 
 SRC_FILES	=	minishell.c \
 				init/init.c \
-				lexer/lexer.c lexer/tokenize.c lexer/copy.c lexer/utils.c lexer/checker.c \
+				lexer/lexer.c lexer/tokenize.c lexer/copy.c lexer/quote.c lexer/utils.c lexer/checker.c \
 				expanser/expanser.c expanser/var.c expanser/quote.c expanser/insert.c \
 				parser/parser.c parser/utils.c parser/path.c \
 				utils/display.c utils/error.c utils/utils.c utils/string.c \
