+++ inc/minishell.h
@@ -11,7 +11,7 @@
 /* ************************************************************************** */
 
 #ifndef MINISHELL_H
-#define MINISHELL_H
+# define MINISHELL_H
 
 // SYSTEM LIBRARIES
 # include <stdio.h>
@@ -28,7 +28,7 @@
 
 // global
 
-extern t_global g_global; 
+extern t_global	g_global;
 
 /*
 **
@@ -49,20 +49,20 @@ int		is_delimiter(char *str);
 int		is_isspace(char c);
 int		is_quote(char *str, int *type);
 int		is_oper(char c1, char c2, int *type);
-int    is_special(char *str, int *type);
+int		is_special(char *str, int *type);
+int		lexer_checker(t_token *head);
 void	add_token(t_token *new, t_token **tokens);
 char	*copy_chars(char *str, int *i, int n);
 char	*copy_word(char *str, int *i);
 char	*copy_var(char *str, int *i);
 char	*copy_quote(char *str, int *i);
-int		lexer_checker(t_token *head);
 /*
 **
 ** EXPANSER
 **
 */
 int		ms_expanser(t_token **token);
-int		var_expanser(t_token **tokens);
+int		expanse_var(t_token **tokens);
 int		expanse_quote(t_token *tokens, char *str, int idx);
 int		*ft_tabint(const int *t, int c);
 void	insert_token(t_token **tokens, t_token *new);
@@ -88,17 +88,18 @@ void	free_tab(char **tab);
 **
 */
 int		get_tablen(char **tab);
+int		*ft_tabint(const int *t, int c);
 int		ft_isdigit(char c);
 int		ft_strcmp(char *s1, char *s2);
 int		ft_strncmp(char *s1, char *s2, int n);
+int		ft_strlen(char *str);
+int		err_msg(int err, char c);
 char	*ft_sjoin(char *s1, char *s2);
 char	*ft_strjoin(char *s1, char *s2);
-int		ft_strlen(char *str);
 char	*ft_strncpy(char *dst, char *src, int n);
-void	display_tokens();
+void	display_tokens(void);
 void	display_cmds(void);
 void	display_specific_tokens(t_token *head);
-int		err_msg(int	err, char c);
 /*
 **
 ** BUILTINS
