+++ inc/structs.h
@@ -10,13 +10,13 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#ifndef STRUCTS_H
+# define STRUCTS_H
 
-#ifndef MS_STRUCTS_H
-#define MS_STRUCTS_H
+# include "minishell.h"
 
 // TOKENS
-typedef struct	s_token
+typedef struct s_token
 {
 	char			*value;
 	int				type;
@@ -28,54 +28,52 @@ typedef struct	s_token
 }				t_token;
 
 // COMMANDS
-typedef struct	s_cmd
+typedef struct s_cmd
 {
 	int				fd_in;
 	int				fd_out;
 	int				builtin;
 	int				redir_in;
 	int				redir_out;
-	char			*cmd; // absolute path
+	char			*cmd;
 	char			**cmd_args;
 	char			**args;
 }				t_cmd;
 
 // CMD LIST
-typedef struct	s_cmdlst
+typedef struct s_cmdlst
 {
-	int 				type;
+	int					type;
 	t_cmd				*cmd;
 	struct s_cmdlst		*next;
 	struct s_cmdlst		*prev;
-	
 }				t_cmdlst;
 
 // TERMINAL -- termios
-typedef struct	s_terminal
+typedef struct s_terminal
 {
 	struct termios	dftl;
 	struct termios	new;
 }				t_terminal;
 
-
 // SHELL -- minishell data
-typedef struct	s_shell
+typedef struct s_shell
 {
 	char		**env;
 	char		*input;
 }				t_shell;
 
 // GLOBAL (variable)
-typedef struct	s_data
+typedef struct s_data
 {
 	int				err;
 	t_shell			shell;
 	t_terminal		terminal;
-	t_cmdlst		*cmds;		// linked list
-	t_token			*tokens;	// linked list
+	t_cmdlst		*cmds;
+	t_token			*tokens;
 }				t_data;
 
-typedef struct	s_global
+typedef struct s_global
 {
 	t_data	*data;
 }			t_global;
